+++ src/main/java/com/mall/dto/common/ErrorResponseDto.java
@@ -1,4 +1,4 @@
-package com.mall.dto.response;
+package com.mall.dto.common;
 
 import com.mall.exception.CustomException;
 import lombok.AllArgsConstructor;
